requirements.txt
llama-index>=0.12.5
llama-index-llms-azure-openai>=0.3.0
python-dotenv>=1.0.1  
python-docx>=1.1.2
